wip: start of ai rewrite of esm_environment, probably nothing works out of the box
