+++ setup.py
@@ -17,6 +17,7 @@ requirements = [
     "PyGithub==1.55",
     "colorama==0.4.5",
     "coloredlogs==15.0.1",  # NOTE(PG): Should be removed during cleanup for loguru instead
+    "dpath",
     "emoji==1.7.0",
     "f90nml==1.4.2",
     "gfw-creator==0.2.2",
@@ -106,12 +107,14 @@ setup(
         "esm_tools.namelists": "namelists",
         "esm_tools.runscripts": "runscripts",
         "esm_tools.couplings": "couplings",
+        "esm_tools.templates": "templates",
     },
     package_data={
         "esm_tools.configs": ["../configs/*"],
         "esm_tools.namelists": ["../namelists/*"],
         "esm_tools.runscripts": ["../runscripts/*"],
         "esm_tools.couplings": ["../couplings/*"],
+        "esm_tools.templates": ["../templates/*"],
     },
     setup_requires=setup_requirements,
     test_suite="tests",
